main/superd-config: new aport
This is a collection of .service files that can be loaded by superd for
starting/managing user services.

The base package installs a superd.desktop file that can be used by UIs
that xdg autostart things, with the idea that they would be pointed to
this file/dir and let superd start everything else.

The subpackages are generated automatically, using the .service file
name(s) to determine the package to "latch" onto. For example, a
"foo.bar.service" file would make a new subpackage that had:
    install_if="foo superd"
    install foo.bar.service "$subpkgdir/usr/share/superd/bar.service"
This file name format is explained in more detail in the APKBUILD.

If a package in Alpine or pmOS already ships a .service file, the
subpackage could install a symlink to that file under /usr/share/superd
so that superd is aware of it. None of the packages added in this
initial implementation ship .service files in the alpine packagea... so
a new, simple .service was added here for now.